fix(#2456): split dataset_generator.py (1150 LOC) into three focused modules
- _dataset_models.py (280 LOC): ParameterRange, ControlProfile, GeneratorConfig,
  SimulationSample, TrainingDataset data classes
- _dataset_export_mixin.py (387 LOC): _DatasetExportMixin with export_to_hdf5,
  export_to_sqlite, export_to_csv, export methods
- dataset_generator.py (549 LOC): DatasetGenerator class inheriting _DatasetExportMixin,
  containing core generate() + simulation helper methods; re-exports all public names

All 6 public names remain importable from dataset_generator (backward compat).
Adds TDD contract tests: structure, LOC budgets, and public API import checks.
