+++ src/shared/python/data_io/dataset_generator.py
@@ -4,6 +4,9 @@ Generates large-scale simulation datasets by varying inputs across physics engin
 Records all kinematics (q, v, a), kinetics (tau, forces, energies), and model
 data (inertia, bias forces, Jacobians) into structured databases for ML training.
 
+Data models are in _dataset_models.py.
+Export methods are in _dataset_export_mixin.py.
+
 Design by Contract:
     Preconditions:
         - Engine must implement PhysicsEngine protocol
@@ -33,11 +36,6 @@ Usage:
 from __future__ import annotations
 
 import contextlib
-import json
-import sqlite3
-import time
-from dataclasses import dataclass, field
-from pathlib import Path
 from typing import Any
 
 import numpy as np
@@ -47,273 +45,24 @@ from src.shared.python.core.error_utils import SimulationError
 from src.shared.python.engine_core.interfaces import PhysicsEngine
 from src.shared.python.logging_pkg.logging_config import get_logger
 
+# Re-export public data models for backward compatibility
+from ._dataset_export_mixin import _DatasetExportMixin
+from ._dataset_models import (
+    ControlProfile,
+    GeneratorConfig,
+    ParameterRange,
+    SimulationSample,
+    TrainingDataset,
+)
+
 logger = get_logger(__name__)
-
-
-@dataclass
-class ParameterRange:
-    """Defines a range for parameter variation.
-
-    Attributes:
-        name: Parameter identifier.
-        min_val: Minimum value.
-        max_val: Maximum value.
-        distribution: Sampling distribution ('uniform', 'normal', 'linspace').
-        num_points: Number of discrete points for linspace distribution.
-    """
-
-    name: str
-    min_val: float
-    max_val: float
-    distribution: str = "uniform"
-    num_points: int = 10
-
-    def __post_init__(self) -> None:
-        """Validate parameter range.
-
-        Raises:
-            ValueError: If min_val > max_val or distribution is unknown.
-        """
-        if self.min_val > self.max_val:
-            raise ValueError(
-                f"Invalid range for '{self.name}': "
-                f"min_val ({self.min_val}) > max_val ({self.max_val})"
-            )
-        valid_distributions = {"uniform", "normal", "linspace"}
-        if self.distribution not in valid_distributions:
-            raise ValueError(
-                f"Unknown distribution '{self.distribution}'. "
-                f"Valid: {sorted(valid_distributions)}"
-            )
-
-    def sample(self, rng: np.random.Generator) -> float:
-        """Sample a value from this range.
-
-        Args:
-            rng: NumPy random generator.
-
-        Returns:
-            Sampled value within the defined range.
-        """
-        if not (rng is not None):
-            raise ValueError("rng must be provided")
-        if not (rng is not None):
-            raise ValueError("rng must be provided")
-        if self.distribution == "uniform":
-            return float(rng.uniform(self.min_val, self.max_val))
-        if self.distribution == "normal":
-            mean = (self.min_val + self.max_val) / 2.0
-            std = (self.max_val - self.min_val) / 6.0  # 99.7% within range
-            val = float(rng.normal(mean, std))
-            return float(np.clip(val, self.min_val, self.max_val))
-        # linspace
-        points = np.linspace(self.min_val, self.max_val, self.num_points)
-        return float(rng.choice(points))
-
-    def linspace(self) -> np.ndarray:
-        """Generate evenly spaced values across the range.
-
-        Returns:
-            Array of evenly spaced values.
-        """
-        return np.linspace(self.min_val, self.max_val, self.num_points)
-
-
-@dataclass
-class ControlProfile:
-    """Defines a control input profile for dataset generation.
-
-    Attributes:
-        name: Profile identifier.
-        profile_type: Type of control profile.
-        parameters: Profile-specific parameters.
-    """
-
-    name: str
-    profile_type: str = "zero"  # zero, constant, sinusoidal, random, step
-    parameters: dict[str, Any] = field(default_factory=dict)
-
-    def generate(
-        self, n_actuators: int, n_steps: int, dt: float, rng: np.random.Generator
-    ) -> np.ndarray:
-        """Generate control input sequence.
-
-        Args:
-            n_actuators: Number of actuators/DOFs.
-            n_steps: Number of timesteps.
-            dt: Timestep size.
-            rng: Random generator.
-
-        Returns:
-            Control array of shape (n_steps, n_actuators).
-        """
-        if not (n_actuators is not None):
-            raise ValueError("n_actuators must be provided")
-        if not (n_actuators is not None):
-            raise ValueError("n_actuators must be provided")
-        if self.profile_type == "zero":
-            return np.zeros((n_steps, n_actuators))
-        if self.profile_type == "constant":
-            magnitude = self.parameters.get("magnitude", 1.0)
-            return np.full((n_steps, n_actuators), magnitude)
-        if self.profile_type == "sinusoidal":
-            freq = self.parameters.get("frequency", 1.0)
-            amplitude = self.parameters.get("amplitude", 1.0)
-            t = np.arange(n_steps) * dt
-            base = amplitude * np.sin(2.0 * np.pi * freq * t)
-            return np.column_stack([base] * n_actuators)
-        if self.profile_type == "random":
-            scale = self.parameters.get("scale", 1.0)
-            return rng.normal(0, scale, (n_steps, n_actuators))
-        if self.profile_type == "step":
-            magnitude = self.parameters.get("magnitude", 1.0)
-            step_time = self.parameters.get("step_time", 0.5)
-            step_idx = int(step_time / dt)
-            profile = np.zeros((n_steps, n_actuators))
-            if step_idx < n_steps:
-                profile[step_idx:] = magnitude
-            return profile
-        return np.zeros((n_steps, n_actuators))
-
-
-@dataclass
-class GeneratorConfig:
-    """Configuration for dataset generation.
-
-    Attributes:
-        num_samples: Number of simulation runs to generate.
-        duration: Duration of each simulation in seconds.
-        timestep: Simulation timestep in seconds.
-        seed: Random seed for reproducibility.
-        vary_initial_positions: Whether to randomize initial joint positions.
-        vary_initial_velocities: Whether to randomize initial joint velocities.
-        position_ranges: Ranges for initial position variation.
-        velocity_ranges: Ranges for initial velocity variation.
-        control_profiles: Control profiles to sample from.
-        record_mass_matrix: Whether to record inertia matrices.
-        record_bias_forces: Whether to record bias forces.
-        record_gravity: Whether to record gravity forces.
-        record_jacobians: Whether to record Jacobians.
-        record_contact_forces: Whether to record contact forces.
-        record_drift_control: Whether to record drift/control decomposition.
-        record_counterfactuals: Whether to record ZTCF/ZVCF.
-        output_fields: Explicit list of fields to record (None = all).
-    """
-
-    num_samples: int = 100
-    duration: float = 2.0
-    timestep: float = 0.002
-    seed: int = 42
-    vary_initial_positions: bool = True
-    vary_initial_velocities: bool = False
-    position_ranges: list[ParameterRange] = field(default_factory=list)
-    velocity_ranges: list[ParameterRange] = field(default_factory=list)
-    control_profiles: list[ControlProfile] = field(
-        default_factory=lambda: [
-            ControlProfile(name="zero"),
-        ]
-    )
-    record_mass_matrix: bool = True
-    record_bias_forces: bool = True
-    record_gravity: bool = True
-    record_jacobians: bool = False
-    record_contact_forces: bool = True
-    record_drift_control: bool = True
-    record_counterfactuals: bool = False
-    output_fields: list[str] | None = None
-
-    def __post_init__(self) -> None:
-        """Validate configuration.
-
-        Raises:
-            ValueError: If configuration values are invalid.
-        """
-        if self.num_samples <= 0:
-            raise ValueError(f"num_samples must be positive, got {self.num_samples}")
-        if self.duration <= 0:
-            raise ValueError(f"duration must be positive, got {self.duration}")
-        if self.timestep <= 0:
-            raise ValueError(f"timestep must be positive, got {self.timestep}")
-        if self.duration < self.timestep:
-            raise ValueError(
-                f"duration ({self.duration}) must be >= timestep ({self.timestep}); "
-                "otherwise no steps would be recorded"
-            )
-
-
-@dataclass
-class SimulationSample:
-    """A single simulation run's recorded data.
-
-    Attributes:
-        sample_id: Unique sample identifier.
-        metadata: Configuration and provenance metadata.
-        times: Time array (n_steps,).
-        positions: Joint positions (n_steps, n_q).
-        velocities: Joint velocities (n_steps, n_v).
-        accelerations: Joint accelerations (n_steps, n_v).
-        torques: Applied joint torques (n_steps, n_v).
-        mass_matrices: Mass matrices per step (n_steps, n_v, n_v) or None.
-        bias_forces: Bias forces per step (n_steps, n_v) or None.
-        gravity_forces: Gravity forces per step (n_steps, n_v) or None.
-        contact_forces: Contact forces per step (n_steps, 3) or None.
-        drift_accelerations: Drift accelerations (n_steps, n_v) or None.
-        control_accelerations: Control accelerations (n_steps, n_v) or None.
-        energies: Energy data dict.
-    """
-
-    sample_id: int
-    metadata: dict[str, Any]
-    times: np.ndarray
-    positions: np.ndarray
-    velocities: np.ndarray
-    accelerations: np.ndarray
-    torques: np.ndarray
-    mass_matrices: np.ndarray | None = None
-    bias_forces: np.ndarray | None = None
-    gravity_forces: np.ndarray | None = None
-    contact_forces: np.ndarray | None = None
-    drift_accelerations: np.ndarray | None = None
-    control_accelerations: np.ndarray | None = None
-    energies: dict[str, np.ndarray] = field(default_factory=dict)
-
-
-@dataclass
-class TrainingDataset:
-    """Collection of simulation samples forming a training dataset.
-
-    Attributes:
-        samples: List of simulation samples.
-        config: Generator configuration used.
-        model_name: Name of the model used.
-        engine_name: Name of the physics engine used.
-        joint_names: Names of joints in the model.
-        creation_time: Unix timestamp of dataset creation.
-    """
-
-    samples: list[SimulationSample]
-    config: GeneratorConfig
-    model_name: str
-    engine_name: str
-    joint_names: list[str]
-    creation_time: float = field(default_factory=time.time)
-
-    @property
-    def num_samples(self) -> int:
-        """Number of samples in the dataset."""
-        return len(self.samples)
-
-    @property
-    def total_frames(self) -> int:
-        """Total number of frames across all samples."""
-        return sum(len(s.times) for s in self.samples)
 
 
 @invariant(
     lambda self: self.engine is not None,
     "DatasetGenerator must have a valid engine reference",
 )
-class DatasetGenerator:
+class DatasetGenerator(_DatasetExportMixin):
     """Generates simulation datasets for neural network training.
 
     Uses a PhysicsEngine to run simulations with varied inputs and records
@@ -789,362 +538,12 @@ class DatasetGenerator:
         n_q, _ = self._get_dimensions()
         return [f"joint_{i}" for i in range(n_q)]
 
-    def export_to_hdf5(self, dataset: TrainingDataset, output_path: str | Path) -> Path:
-        """Export dataset to HDF5 format.
 
-        Args:
-            dataset: Training dataset to export.
-            output_path: Output file path (without extension).
-
-        Returns:
-            Path to the created HDF5 file.
-
-        Raises:
-            ImportError: If h5py is not available.
-        """
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        try:
-            import h5py
-        except ImportError:
-            raise ImportError(
-                "h5py required for HDF5 export: pip install h5py"
-            ) from None
-
-        output_path = Path(output_path)
-        if not output_path.suffix:
-            output_path = output_path.with_suffix(".hdf5")
-
-        output_path.parent.mkdir(parents=True, exist_ok=True)
-
-        with h5py.File(str(output_path), "w") as f:
-            self._write_hdf5_metadata(f, dataset)
-            samples_grp = f.create_group("samples")
-            for sample in dataset.samples:
-                self._write_hdf5_sample(samples_grp, sample)
-
-        logger.info("Exported dataset to HDF5: %s", output_path)
-        return output_path
-
-    @staticmethod
-    def _write_hdf5_metadata(f: Any, dataset: TrainingDataset) -> None:
-        """Write dataset-level metadata to an HDF5 file."""
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        meta = f.create_group("metadata")
-        meta.attrs["model_name"] = dataset.model_name
-        meta.attrs["engine_name"] = dataset.engine_name
-        meta.attrs["num_samples"] = dataset.num_samples
-        meta.attrs["total_frames"] = dataset.total_frames
-        meta.attrs["creation_time"] = dataset.creation_time
-        meta.attrs["duration"] = dataset.config.duration
-        meta.attrs["timestep"] = dataset.config.timestep
-        meta.attrs["seed"] = dataset.config.seed
-
-        if dataset.joint_names:
-            meta.create_dataset(
-                "joint_names",
-                data=[n.encode("utf-8") for n in dataset.joint_names],
-            )
-
-    @staticmethod
-    def _write_hdf5_sample(samples_grp: Any, sample: SimulationSample) -> None:
-        """Write a single sample's data to an HDF5 samples group."""
-        if not (sample is not None):
-            raise ValueError("sample must be provided")
-        if not (sample is not None):
-            raise ValueError("sample must be provided")
-        s_grp = samples_grp.create_group(f"sample_{sample.sample_id:06d}")
-        s_grp.create_dataset("times", data=sample.times, compression="gzip")
-        s_grp.create_dataset("positions", data=sample.positions, compression="gzip")
-        s_grp.create_dataset("velocities", data=sample.velocities, compression="gzip")
-        s_grp.create_dataset(
-            "accelerations", data=sample.accelerations, compression="gzip"
-        )
-        s_grp.create_dataset("torques", data=sample.torques, compression="gzip")
-
-        optional_fields = [
-            ("mass_matrices", sample.mass_matrices),
-            ("bias_forces", sample.bias_forces),
-            ("gravity_forces", sample.gravity_forces),
-            ("contact_forces", sample.contact_forces),
-            ("drift_accelerations", sample.drift_accelerations),
-            ("control_accelerations", sample.control_accelerations),
-        ]
-        for field_name, field_data in optional_fields:
-            if field_data is not None:
-                s_grp.create_dataset(field_name, data=field_data, compression="gzip")
-
-        if sample.energies:
-            e_grp = s_grp.create_group("energies")
-            for key, arr in sample.energies.items():
-                e_grp.create_dataset(key, data=arr, compression="gzip")
-
-        s_grp.attrs["metadata"] = json.dumps(sample.metadata)
-
-    def export_to_sqlite(
-        self, dataset: TrainingDataset, output_path: str | Path
-    ) -> Path:
-        """Export dataset to SQLite database.
-
-        Args:
-            dataset: Training dataset to export.
-            output_path: Output database path.
-
-        Returns:
-            Path to the created SQLite database.
-        """
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        output_path = Path(output_path)
-        if not output_path.suffix:
-            output_path = output_path.with_suffix(".db")
-
-        output_path.parent.mkdir(parents=True, exist_ok=True)
-
-        conn = sqlite3.connect(str(output_path))
-        try:
-            cursor = conn.cursor()
-            self._create_sqlite_tables(cursor)
-            self._insert_sqlite_metadata(cursor, dataset)
-            for sample in dataset.samples:
-                self._insert_sqlite_sample(cursor, sample)
-            conn.commit()
-        finally:
-            conn.close()
-
-        logger.info("Exported dataset to SQLite: %s", output_path)
-        return output_path
-
-    @staticmethod
-    def _create_sqlite_tables(cursor: sqlite3.Cursor) -> None:
-        """Create the SQLite schema tables."""
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS dataset_metadata (
-                key TEXT PRIMARY KEY,
-                value TEXT
-            )
-        """
-        )
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS samples (
-                sample_id INTEGER PRIMARY KEY,
-                metadata_json TEXT,
-                n_steps INTEGER,
-                n_q INTEGER,
-                n_v INTEGER
-            )
-        """
-        )
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS frames (
-                sample_id INTEGER,
-                step INTEGER,
-                time REAL,
-                positions_json TEXT,
-                velocities_json TEXT,
-                accelerations_json TEXT,
-                torques_json TEXT,
-                kinetic_energy REAL,
-                PRIMARY KEY (sample_id, step),
-                FOREIGN KEY (sample_id) REFERENCES samples(sample_id)
-            )
-        """
-        )
-
-    @staticmethod
-    def _insert_sqlite_metadata(
-        cursor: sqlite3.Cursor, dataset: TrainingDataset
-    ) -> None:
-        """Insert dataset-level metadata into the SQLite database."""
-        if not (cursor is not None):
-            raise ValueError("cursor must be provided")
-        if not (cursor is not None):
-            raise ValueError("cursor must be provided")
-        meta_items = [
-            ("model_name", dataset.model_name),
-            ("engine_name", dataset.engine_name),
-            ("num_samples", str(dataset.num_samples)),
-            ("total_frames", str(dataset.total_frames)),
-            ("creation_time", str(dataset.creation_time)),
-            ("seed", str(dataset.config.seed)),
-            ("duration", str(dataset.config.duration)),
-            ("timestep", str(dataset.config.timestep)),
-            ("joint_names", json.dumps(dataset.joint_names)),
-        ]
-        cursor.executemany(
-            "INSERT OR REPLACE INTO dataset_metadata (key, value) VALUES (?, ?)",
-            meta_items,
-        )
-
-    @staticmethod
-    def _insert_sqlite_sample(cursor: sqlite3.Cursor, sample: SimulationSample) -> None:
-        """Insert a single sample and its frames into the SQLite database."""
-        if not (cursor is not None):
-            raise ValueError("cursor must be provided")
-        if not (cursor is not None):
-            raise ValueError("cursor must be provided")
-        n_steps = len(sample.times)
-        n_q = sample.positions.shape[1] if sample.positions.ndim > 1 else 0
-        n_v = sample.velocities.shape[1] if sample.velocities.ndim > 1 else 0
-
-        cursor.execute(
-            "INSERT INTO samples (sample_id, metadata_json, n_steps, n_q, n_v) "
-            "VALUES (?, ?, ?, ?, ?)",
-            (sample.sample_id, json.dumps(sample.metadata), n_steps, n_q, n_v),
-        )
-
-        frame_rows = []
-        for step in range(n_steps):
-            ke = (
-                float(sample.energies["kinetic"][step])
-                if "kinetic" in sample.energies
-                else 0.0
-            )
-            frame_rows.append(
-                (
-                    sample.sample_id,
-                    step,
-                    float(sample.times[step]),
-                    json.dumps(sample.positions[step].tolist()),
-                    json.dumps(sample.velocities[step].tolist()),
-                    json.dumps(sample.accelerations[step].tolist()),
-                    json.dumps(sample.torques[step].tolist()),
-                    ke,
-                )
-            )
-
-        cursor.executemany(
-            "INSERT INTO frames "
-            "(sample_id, step, time, positions_json, velocities_json, "
-            "accelerations_json, torques_json, kinetic_energy) "
-            "VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-            frame_rows,
-        )
-
-    def export_to_csv(self, dataset: TrainingDataset, output_dir: str | Path) -> Path:
-        """Export dataset to CSV files (one per sample).
-
-        Args:
-            dataset: Training dataset to export.
-            output_dir: Output directory for CSV files.
-
-        Returns:
-            Path to the output directory.
-        """
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        if not (dataset is not None):
-            raise ValueError("dataset must be provided")
-        output_dir = Path(output_dir)
-        output_dir.mkdir(parents=True, exist_ok=True)
-
-        for sample in dataset.samples:
-            n_steps = len(sample.times)
-            n_q = sample.positions.shape[1]
-            n_v = sample.velocities.shape[1]
-
-            # Build header
-            headers = ["time"]
-            headers.extend([f"q_{i}" for i in range(n_q)])
-            headers.extend([f"v_{i}" for i in range(n_v)])
-            headers.extend([f"a_{i}" for i in range(n_v)])
-            headers.extend([f"tau_{i}" for i in range(n_v)])
-            headers.append("kinetic_energy")
-
-            # Build data matrix
-            data_cols = [sample.times.reshape(-1, 1)]
-            data_cols.append(sample.positions)
-            data_cols.append(sample.velocities)
-            data_cols.append(sample.accelerations)
-            data_cols.append(sample.torques)
-
-            ke = sample.energies.get("kinetic", np.zeros(n_steps))
-            data_cols.append(ke.reshape(-1, 1))
-
-            data = np.hstack(data_cols)
-
-            filepath = output_dir / f"sample_{sample.sample_id:06d}.csv"
-            np.savetxt(
-                str(filepath),
-                data,
-                delimiter=",",
-                header=",".join(headers),
-                comments="",
-            )
-
-        # Write metadata file
-        meta_path = output_dir / "metadata.json"
-        meta = {
-            "model_name": dataset.model_name,
-            "engine_name": dataset.engine_name,
-            "num_samples": dataset.num_samples,
-            "total_frames": dataset.total_frames,
-            "joint_names": dataset.joint_names,
-            "config": {
-                "duration": dataset.config.duration,
-                "timestep": dataset.config.timestep,
-                "seed": dataset.config.seed,
-                "num_samples": dataset.config.num_samples,
-            },
-        }
-        with open(meta_path, "w") as f:
-            json.dump(meta, f, indent=2)
-
-        logger.info("Exported dataset to CSV: %s", output_dir)
-        return output_dir
-
-    @precondition(
-        lambda self, dataset, output_path, format="hdf5": dataset is not None,
-        "Dataset must not be None",
-    )
-    @precondition(
-        lambda self, dataset, output_path, format="hdf5": (
-            output_path is not None and len(str(output_path)) > 0
-        ),
-        "Output path must be a non-empty string or Path",
-    )
-    @precondition(
-        lambda self, dataset, output_path, format="hdf5": (
-            format in ("hdf5", "sqlite", "db", "csv")
-        ),
-        "Export format must be one of: hdf5, sqlite, db, csv",
-    )
-    def export(
-        self,
-        dataset: TrainingDataset,
-        output_path: str | Path,
-        format: str = "hdf5",
-    ) -> Path:
-        """Export dataset in the specified format.
-
-        Args:
-            dataset: Training dataset to export.
-            output_path: Output path (file or directory depending on format).
-            format: Export format ('hdf5', 'sqlite', 'csv').
-
-        Returns:
-            Path to the exported data.
-
-        Raises:
-            ValueError: If format is not supported.
-        """
-        format = format.lower()
-        if format == "hdf5":
-            return self.export_to_hdf5(dataset, output_path)
-        if format in ("sqlite", "db"):
-            return self.export_to_sqlite(dataset, output_path)
-        if format == "csv":
-            return self.export_to_csv(dataset, output_path)
-        raise ValueError(
-            f"Unsupported export format: {format}. Supported: hdf5, sqlite, csv"
-        )
+__all__ = [
+    "ControlProfile",
+    "DatasetGenerator",
+    "GeneratorConfig",
+    "ParameterRange",
+    "SimulationSample",
+    "TrainingDataset",
+]
